src/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.9.0
attrs==25.3.0
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
crispy-bootstrap5==2025.6
dj-rest-auth==7.0.1
Django==5.2.4
django-allauth==65.9.0
django-crispy-forms==2.4
django-phonenumber-field==8.1.0
djangorestframework==3.16.0
drf-spectacular==0.28.0
environs==14.2.0
ff3==1.0.2
flake8==7.3.0
inflection==0.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
marshmallow==4.0.0
mccabe==0.7.0
packaging==25.0
phonenumbers==9.0.8
pillow==11.3.0
prompt_toolkit==3.0.51
psycopg==3.2.9
psycopg-binary==3.2.9
pycodestyle==2.14.0
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
pyflakes==3.4.0
pyngo==2.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
PyYAML==6.0.2
redis==6.2.0
referencing==0.36.2
rpds-py==0.26.0
setuptools==65.5.1
six==1.17.0
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
vine==5.1.0
wcwidth==0.2.13
wheel==0.38.4
